refactor: externalise user prompts
