+++ src/utils/extensions.ts
@@ -6,6 +6,7 @@ import * as fs from "fs";
 import * as path from "path";
 import * as yaml from "js-yaml";
 import { installQuartoExtension } from "./quarto";
+import { askTrustAuthors, askConfirmInstall } from "./ask";
 
 interface ExtensionQuickPickItem extends vscode.QuickPickItem {
 	url?: string;
@@ -84,51 +85,8 @@ export async function installQuartoExtensions(
 	}
 	const mutableSelectedExtensions: ExtensionQuickPickItem[] = [...selectedExtensions];
 
-	const config = vscode.workspace.getConfiguration("quartoWizard.ask");
-	let configTrustAuthors = config.get<string>("trustAuthors");
-	let configConfirmInstall = config.get<string>("confirmInstall");
-
-	if (configTrustAuthors === "always") {
-		const trustAuthors = await vscode.window.showQuickPick(
-			[
-				{ label: "Yes", description: "Trust authors." },
-				{ label: "No", description: "Do not trust authors." },
-				{ label: "Never ask again", description: "Change setting to never ask again." },
-			],
-			{
-				placeHolder: "Do you trust the authors of the selected extension(s)?",
-			}
-		);
-		if (trustAuthors?.label === "Never ask again") {
-			await config.update("trustAuthors", "never", vscode.ConfigurationTarget.Global);
-		} else if (trustAuthors?.label !== "Yes") {
-			const message = "Operation cancelled because the authors are not trusted.";
-			log.appendLine(message);
-			vscode.window.showInformationMessage(message);
-			return;
-		}
-	}
-
-	if (configConfirmInstall === "always") {
-		const installWorkspace = await vscode.window.showQuickPick(
-			[
-				{ label: "Yes", description: "Install extensions." },
-				{ label: "No", description: "Do not install extensions." },
-				{ label: "Never ask again", description: "Change setting to never ask again." },
-			],
-			{
-				placeHolder: "Do you want to install the selected extension(s)?",
-			}
-		);
-		if (installWorkspace?.label === "Never ask again") {
-			await config.update("confirmInstall", "never", vscode.ConfigurationTarget.Global);
-		} else if (installWorkspace?.label !== "Yes") {
-			const message = "Operation cancelled by the user.";
-			log.appendLine(message);
-			vscode.window.showInformationMessage(message);
-			return;
-		}
-	}
+	if ((await askTrustAuthors(log)) !== 0) return;
+	if ((await askConfirmInstall(log)) !== 0) return;
 
 	vscode.window.withProgress(
 		{
